Make ModuleParser abstract.
